requirements.txt

absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
albucore==0.0.23
albumentations==1.4.24
alembic==1.14.0
annotated-types==0.7.0
attrs==24.2.0
batchgenerators==0.21
blinker==1.9.0
blosc==1.11.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.39.0
Deprecated==1.2.15
docker==7.1.0
einops==0.8.0
eval_type_backport==0.2.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
huggingface-hub==0.26.5
humanize==4.11.0
hydra-core==1.3.2
idna==3.10
imageio==2.36.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.9
linecache2==1.0.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mlflow==2.19.0
mlflow-skinny==2.19.0
monai==1.4.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
numcodecs
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pydicom==3.0.1
pyparsing==3.2.0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
shellingham==1.5.4
SimpleITK==2.4.0
simsimd==6.2.1
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
stringzilla==3.11.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.12.12
timm==1.0.12
torch==2.2.2
torchio==0.20.3
torchmetrics==1.6.0
torchvision==0.17.2
tqdm==4.67.1
traceback2==1.4.0
triton==2.2.0
typer==0.15.1
tzdata==2024.2
unittest2==1.1.0
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
